# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pragati Bharti Document Intelligence API",
    "description": "Production-grade, scalable document intelligence and question extraction service for exam and question-bank material. Converts arbitrary PDF and image layouts into structured, machine-readable question banks with answers, bounding boxes, and confidence metrics.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/documents": {
      "post": {
        "tags": [
          "Documents"
        ],
        "summary": "Upload Document",
        "description": "Upload an exam PDF or image (JPG/PNG).\nReturns immediately with HTTP 202 Accepted and a document ID for asynchronous background processing.",
        "operationId": "upload_document_api_v1_documents_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_api_v1_documents_post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "List Documents",
        "description": "Retrieve a paginated list of uploaded documents for the current user.",
        "operationId": "list_documents_api_v1_documents_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_DocumentResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/batch": {
      "post": {
        "tags": [
          "Documents"
        ],
        "summary": "Upload Documents Batch",
        "description": "Upload multiple files simultaneously and optionally auto-group them.",
        "operationId": "upload_documents_batch_api_v1_documents_batch_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_documents_batch_api_v1_documents_batch_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/documents/{document_id}": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "Get Document",
        "description": "Retrieve metadata, current status, and stage timings for a specific document.",
        "operationId": "get_document_api_v1_documents__document_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Documents"
        ],
        "summary": "Delete Document",
        "description": "Delete a document, its extracted pages, questions, and storage objects.",
        "operationId": "delete_document_api_v1_documents__document_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}/status": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "Get Document Status",
        "description": "Lightweight polling endpoint returning real-time progress percentage and current stage.",
        "operationId": "get_document_status_api_v1_documents__document_id__status_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}/reprocess": {
      "post": {
        "tags": [
          "Documents"
        ],
        "summary": "Reprocess Document",
        "description": "Re-trigger the processing pipeline for an existing document.",
        "operationId": "reprocess_document_api_v1_documents__document_id__reprocess_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}/pages": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "Get Document Pages",
        "description": "Retrieve details, quality scores, and OCR confidence for all pages in the document.",
        "operationId": "get_document_pages_api_v1_documents__document_id__pages_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentPageResponse"
                  },
                  "title": "Response Get Document Pages Api V1 Documents  Document Id  Pages Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}/pages/{page_no}/image": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "Get Page Image",
        "description": "Stream authorized page preview image for verification and review.",
        "operationId": "get_page_image_api_v1_documents__document_id__pages__page_no__image_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          },
          {
            "name": "page_no",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Page No"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/groups/{group_id}/documents": {
      "post": {
        "tags": [
          "Document Groups"
        ],
        "summary": "Attach Document To Group",
        "description": "Attach a document to a group with a specific role (question_paper, answer_key) and trigger re-association.",
        "operationId": "attach_document_to_group_api_v1_groups__group_id__documents_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Group Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttachDocumentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/groups/{group_id}/documents/{document_id}": {
      "delete": {
        "tags": [
          "Document Groups"
        ],
        "summary": "Detach Document From Group",
        "description": "Remove a document from a group.",
        "operationId": "detach_document_from_group_api_v1_groups__group_id__documents__document_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Group Id"
            }
          },
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AttachDocumentRequest": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "description": "question_paper, answer_key, mixed, unknown",
            "default": "unknown"
          }
        },
        "type": "object",
        "required": [
          "document_id"
        ],
        "title": "AttachDocumentRequest"
      },
      "Body_upload_document_api_v1_documents_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "group_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Id"
          },
          "role_in_group": {
            "type": "string",
            "title": "Role In Group",
            "default": "unknown"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_document_api_v1_documents_post"
      },
      "Body_upload_documents_batch_api_v1_documents_batch_post": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Files"
          },
          "group_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Name"
          }
        },
        "type": "object",
        "required": [
          "files"
        ],
        "title": "Body_upload_documents_batch_api_v1_documents_batch_post"
      },
...
      "DocumentPageResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "page_no": {
            "type": "integer",
            "title": "Page No"
          },
          "rotation_applied": {
            "type": "integer",
            "title": "Rotation Applied"
          },
          "is_scanned": {
            "type": "boolean",
            "title": "Is Scanned"
          },
          "quality_score": {
            "type": "number",
            "title": "Quality Score"
          },
          "ocr_mean_confidence": {
            "type": "number",
            "title": "Ocr Mean Confidence"
          },
          "text_source": {
            "type": "string",
            "title": "Text Source"
          },
          "raw_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Text"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "warnings": {
            "items": {},
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "id",
          "document_id",
          "page_no",
          "rotation_applied",
          "is_scanned",
          "quality_score",
          "ocr_mean_confidence",
          "text_source"
        ],
        "title": "DocumentPageResponse"
      },
      "DocumentResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "owner_id": {
            "type": "string",
            "title": "Owner Id"
          },
          "group_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Id"
          },
          "role_in_group": {
            "type": "string",
            "title": "Role In Group"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "mime_type": {
            "type": "string",
            "title": "Mime Type"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "page_count": {
            "type": "integer",
            "title": "Page Count"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "progress_pct": {
            "type": "integer",
            "title": "Progress Pct"
          },
          "current_stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Stage"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "timings": {
            "additionalProperties": true,
            "type": "object",
            "title": "Timings"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "owner_id",
          "role_in_group",
          "original_filename",
          "mime_type",
          "size_bytes",
          "sha256",
          "page_count",
          "status",
          "progress_pct",
          "created_at"
        ],
        "title": "DocumentResponse"
      },
      "DocumentStatusResponse": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "current_stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Stage"
          },
          "progress_pct": {
            "type": "integer",
            "title": "Progress Pct"
          },
          "timings": {
            "additionalProperties": true,
            "type": "object",
            "title": "Timings"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "status",
          "progress_pct"
        ],
        "title": "DocumentStatusResponse"
      },
      "DocumentUploadResponse": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Document uploaded successfully and queued for processing"
          },
          "duplicate": {
            "type": "boolean",
            "title": "Duplicate",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "status"
        ],
        "title": "DocumentUploadResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedResponse_DocumentResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DocumentResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        },
        "type": "object",
        "required": [
          "items",
          "pagination"
        ],
        "title": "PaginatedResponse[DocumentResponse]"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}